Added a new load balancing scheme that only moves grids to processors on the same node.  This should reduce overall communication and load balancing woes when one has a slow interconnect (i.e. ranger).
Set LoadBalancing = 2 for clusters that schedule processors in blocks
(e.g. ranger), and = 3 for round-robin scheduling.  It tries to
balance the root grids initially by looking at the number of level-1
cells contained in that cell.  It then load-balances the root grids
every LoadBalancingCycleSkip cycles.  There could be unforeseen
problems (as always) because this is brand new...

I will add more details on the trac site later.

--HG--
branch : week-of-code