Initial checkin of the C routines for doing positioning-related calculations, useful for systems where Fortran and C do not agree on the size of the FLOAT vairable.  (See RH's email to enzo-users-l from some time ago for more information.)
PFORTRAN_NAME has now been defined; in cases where PFLOAT is 16, it will use
the C routines.  Otherwise, it will default to the Fortran routines.

Currently will not compile, and I am attempting to track this down.  It seems
likely to be related to some of the type changes in the Enzo-1.5 codebase.

Removed ComputeRT_Timestep, as per John's approval.

--HG--
branch : grid-positioning