MAX_FLAGGING_METHODS was = 7 ! So did you you use any method larger than that this version of the code would have not actually used it ...
There was anothre bug in TurbulenceInitializeGrid that shifted the
velocity by one index which lead to one unpretty edge.

The worries about the Potential turned out to be unfounded.
However, its clear that one can set up problems in such a silly
way that the code cannot possibly get the correct potential ...

--HG--
branch : week-of-code