Cleaned up Grid_Shine.  Removed all of the ONE_SOURCE define garbage. I've been meaning to do this for a long time.
We no longer determine how many energy bins to use here.  This is
taken from RadiationSource->EnergyBins now.  This is determined in
StarParticleRadTransfer.C for Star particles and ReadPhotonSource.C
for static sources.

Added a parameter to turn on helium ionization for Pop III stars.
Before I had to fiddle with the source!

--HG--
branch : week-of-code