This is a simple script that attempts to parse global_data.h in a simple way and write accessor functions for all the variables it knows how to handle.
(This is not a substitute for an overhaul of the configuration system.)

Now the exposed "enzo" module in inline yt will have the function
get_parameter(something), which when called will attempt to return the
parameter if it is known.  It is possible that we could do this faster if we
pre-populated a dictionary with the variables we know will not be changing over
the course of the simulation; I did not do that, however.

--HG--
branch : new_hm_background