Two-step commit.  This step is to add the very initial libconfig modifications. This is a long way from completion; this is just the start.
CanonicalConfiguration.ecfg is designed to be the location of all the default
global values.  Future iterations of this will see it included verbatim in the
executable and read from a string in libconfig.

This is not yet functional, but it is appropriately ifdef'd out.
CanonicalConfiguration.ecfg is not yet a full transcription of a parameter
file.  I will not be returning to this for a while.

--HG--
branch : week-of-code