Fixes to get python_bridge/* working on Triton using the intel compilers.  For what it's worth, the build of the yt stack was fine except that I needed to add:
export CC=icc
export CFLAGS=-fPIC
export F90=ifort
export CXX=icpc
export CXXFLAGS=-fPIC

at the top of the file and --with-shared to the Python 2.6.3 configure line.
I then had to manually remove libpython2.6.a, to force shared library linking.
(This may be a crude solution to a subtle problem.)

--HG--
branch : week-of-code