Modified so that the results are no longer dependent on the order of the fields in the BaryonField array. Previously, if the corrected density or energy (total or internal) was <= 0 in a cell, then the correction for that field would be undone and all further corrections would be set to zero. However, any past corrections would remain as they were. In a simulation where the corrections were to be undone because, for example,the total energy became negative, the fields that were undone would depend on where in the BaryonField array the offending field (e.g. total energy) was located.
The modified version will undo corrections for all fields in an offending cell.

Also, in the scenario I was investigating, I found that the correction for velocity in offending cells might be far larger than the pre-correction value, resulting in v_pre + v_correction - v_correction != v_pre, when the correction was undone. While velocity was the issue for me, this could theoretically occur for any field, and so again, the result would depend on which fields were before and after the total energy in the BaryonField array. To minimize this, density, total energy, and internal energy are now corrected before any other field. (NB: There were some conditions under which the energy fields would not be corrected. Those conditions have been preserved in the modified version.)

--HG--
branch : week-of-code